[VL] Optimize failure output for WrappedVeloxMemoryPool (#1335)
