+++ cpp/core/memory/ArrowMemoryPool.cc
@@ -28,4 +28,31 @@ std::shared_ptr<arrow::MemoryPool> GetDefaultWrappedArrowMemoryPool() {
   return static_pool;
 }
 
+arrow::Status WrappedArrowMemoryPool::Allocate(int64_t size, int64_t alignment, uint8_t** out) {
+  if (!allocator_->AllocateAligned(alignment, size, reinterpret_cast<void**>(out))) {
+    return arrow::Status::Invalid("WrappedMemoryPool: Error allocating " + std::to_string(size) + " bytes");
+  }
+  return arrow::Status::OK();
+}
+
+arrow::Status WrappedArrowMemoryPool::Reallocate(int64_t old_size, int64_t new_size, int64_t alignment, uint8_t** ptr) {
+  if (!allocator_->ReallocateAligned(*ptr, alignment, old_size, new_size, reinterpret_cast<void**>(ptr))) {
+    return arrow::Status::Invalid("WrappedMemoryPool: Error reallocating " + std::to_string(new_size) + " bytes");
+  }
+  return arrow::Status::OK();
+}
+
+void WrappedArrowMemoryPool::Free(uint8_t* buffer, int64_t size, int64_t alignment) {
+  allocator_->Free(buffer, size);
+}
+
+int64_t WrappedArrowMemoryPool::bytes_allocated() const {
+  // fixme use self accountant
+  return allocator_->GetBytes();
+}
+
+std::string WrappedArrowMemoryPool::backend_name() const {
+  return "gluten allocator";
+}
+
 } // namespace gluten
