+++ cpp/core/memory/ArrowMemoryPool.h
@@ -25,4 +25,22 @@ std::shared_ptr<arrow::MemoryPool> AsWrappedArrowMemoryPool(MemoryAllocator* all
 
 std::shared_ptr<arrow::MemoryPool> GetDefaultWrappedArrowMemoryPool();
 
+class WrappedArrowMemoryPool final : public arrow::MemoryPool {
+ public:
+  explicit WrappedArrowMemoryPool(MemoryAllocator* allocator) : allocator_(allocator) {}
+
+  arrow::Status Allocate(int64_t size, int64_t alignment, uint8_t** out) override;
+
+  arrow::Status Reallocate(int64_t old_size, int64_t new_size, int64_t alignment, uint8_t** ptr) override;
+
+  void Free(uint8_t* buffer, int64_t size, int64_t alignment) override;
+
+  int64_t bytes_allocated() const override;
+
+  std::string backend_name() const override;
+
+ private:
+  MemoryAllocator* allocator_;
+};
+
 } // namespace gluten
