+++ cpp/velox/compute/VeloxInitializer.h
@@ -30,6 +30,9 @@ namespace gluten {
 
 /// As a static instance in per executor, initialized at executor startup.
 /// Should not put heavily work here.
+///
+/// FIXME The initializer will be initialized more than once if running Spark
+///   in standalone mode
 class VeloxInitializer {
  public:
   explicit VeloxInitializer(std::unordered_map<std::string, std::string>& conf) {
