+++ cpp/velox/jni/JniWrapper.cc
@@ -44,6 +44,8 @@ jint JNI_OnLoad(JavaVM* vm, void* reserved) {
   if (vm->GetEnv(reinterpret_cast<void**>(&env), JNI_VERSION) != JNI_OK) {
     return JNI_ERR;
   }
+  // logging
+  google::InitGoogleLogging("gluten");
   gluten::GetJniErrorsState()->Initialize(env);
 #ifdef GLUTEN_PRINT_DEBUG
   std::cout << "Loaded Velox backend." << std::endl;
